* 'feature/circle' of https://github.com/zhiyicx/thinksns-plus-android: (58 commits)
  (#1324)   code    TimeOutTaskManager  兼容处理
  (#1324)   code   头像上传兼容 null
  (#1324)  code   兼容 metadata.getDescription()  null
  (#1324)   code  修改部分手机intent数据过大的问题
  (#1324)   code    修改 TSFragment  onDestroy
  (#1324)   code  修复资讯分类为空的兼容
  build v1.3.0
  (#1324)  code  优化音乐内存释放
  (#1324)  code   注册规则增加 空判断
  (#1408)   code   fix  this
  (#1405) code fix this
  (#1404) code fix this
  (#1390) code fix this
  build v1.2.9-beta
  ( #1402)   code   fix  this
  ( #1403)  code   compatible  android 5.1 the status view
  (#1324) code 问题详情的答案列表链接处理
  (#1400) code fix this
  (#1399)   code   排行榜回答显示数量修改
  (#1399)   code  fix  this
  ...

+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/source/local/AnswerDraftBeanGreenDaoImpl.java
@@ -77,6 +77,9 @@ public class AnswerDraftBeanGreenDaoImpl extends CommonCacheImpl<AnswerDraftBean
 
     @Override
     public void deleteSingleCache(AnswerDraftBean dta) {
+        if (dta == null) {
+            return;
+        }
         mAnswerDraftBeanDao.delete(dta);
     }
